requirements.txt
gspread>=6.0.0
google-auth>=2.28.0
playwright>=1.42.0
beautifulsoup4>=4.12.0
requests>=2.31.0
anthropic>=0.40.0
streamlit>=1.42.0
authlib>=1.3.2
